refactor: 컴포넌트 분리, style 파일 분리 - 컴포넌트 style, page style 분리
