+++ src/components/Course/CourseShareCard.tsx
@@ -1,0 +1,73 @@
+import Card from "../Card";
+import Map from "../Map";
+import { Route } from "../../apis/map/types";
+import * as S from "./Course.style";
+import { Button } from "../Button";
+import { useNavigate } from "react-router-dom";
+
+export default function CourseCard({ dailyRoutes }:{
+    dailyRoutes: Route | null | undefined,
+}) {
+    const navigate = useNavigate();
+
+    if(!dailyRoutes) return (
+        <Card>
+            <Card.Title>
+                {"코스가 존재하지 않습니다."}
+            </Card.Title>
+        </Card>
+    );
+    
+    return (
+        <S.StyledCard>
+            <Card.Image>
+                <img src={"https://images.unsplash.com/photo-1487958449943-2429e8be8625?q=80&w=1170&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"} 
+                alt="course" 
+                style={{width:'100%', height:'100%',objectFit:'cover',borderRadius:'1.5rem'}}
+                />
+            </Card.Image>
+
+            <Card.Title>
+                {"저번에 말한 전주 한옥마을에서 가고 싶은곳 있는 사람은 공유하기///"}
+            </Card.Title>
+
+            <Card.Divider/>
+
+            <Card.Item>
+                <S.TitleWrapper>
+                    <S.Caption>{"장소"}</S.Caption>
+
+                    <>
+                        <S.Content>{"전주 한옥마을"}</S.Content>
+                    </>
+                </S.TitleWrapper>
+            </Card.Item>
+
+            <Card.Divider/>
+
+            <Card.Item>
+                <S.TitleWrapper>
+                    <S.Caption>{"기간"}</S.Caption>
+
+                    <>
+                        <S.Content>{"1박 2일 / 날짜 미정"}</S.Content>
+                    </>
+                </S.TitleWrapper>
+            </Card.Item>
+
+            <Card.Item>
+                <Button 
+                    style={{marginTop:'1.125rem'}} 
+                    size="large" 
+                    color="primary"
+                    onClick={() => {
+                        navigate(`/course?courseId=${dailyRoutes.trans_id}`);
+                    }}
+                >
+                    {"원하는 장소 공유하기"}
+                </Button>
+            </Card.Item>
+
+        </S.StyledCard>
+    );
+}
